Move updatable resource types to /models
