+++ src/client.spec.ts
@@ -31,7 +31,6 @@ import {
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
-	UpdatableResource,
 	UpdatableSavedQuery,
 	UpdatableScheduledQuery,
 	UpdatableScheduledScript,
@@ -88,6 +87,7 @@ import {
 	UpdatableMacro,
 	UpdatableNotification,
 	UpdatablePlaybook,
+	UpdatableResource,
 	UploadableAutoExtractor,
 	User,
 	UserPreferences,
