+++ src/models/resource/index.ts
@@ -11,6 +11,9 @@ export * from './is-blank-raw-resource';
 export * from './is-resource';
 export * from './raw-creatable-resource';
 export * from './raw-resource';
+export * from './raw-updatable-resource-metadata';
 export * from './resource';
 export * from './to-raw-creatable-resource';
+export * from './to-raw-updatable-resource-metadata';
 export * from './to-resource';
+export * from './updatable-resource';
